load-bhavcopy-data/requirements.txt
# Function dependencies, for example:
# package>=version
wget
google-cloud-storage
google-cloud-datastore

